Tweaks the JS setup
