+++ resources/js/app.js
@@ -1,10 +1,2 @@
-import 'bootstrap';
-
-import 'elements/turbo-echo-stream-tag';
-import 'elements/turbo-livewire-stream-tag';
-
-import 'libs/stimulus';
-import 'libs/alpine';
-import 'libs/trix';
-import 'libs/turbo';
-import 'libs/bridge';
+import 'libs';
+import 'elements';
