--- resources/js/bootstrap.js
@@ -1,29 +1,0 @@
-import _ from 'lodash';
-window._ = _;
-
-/**
- * We'll load the axios HTTP library which allows us to easily issue requests
- * to our Laravel back-end. This library automatically handles sending the
- * CSRF token as a header based on the value of the "XSRF" token cookie.
- */
-
-import axios from 'axios';
-window.axios = axios;
-
-window.axios.defaults.headers.common['X-Requested-With'] = 'XMLHttpRequest';
-
-import Echo  from 'echo';
-
-window.Echo = Echo;
-
-let token = document.head.querySelector('meta[name="csrf-token"]');
-
-document.addEventListener('turbo:before-fetch-request', (e) => {
-    e.detail.fetchOptions.headers['X-Socket-ID'] = Echo.socketId();
-
-    if (token) {
-        e.detail.fetchOptions.headers['X-CSRF-Token'] = token.content;
-    } else {
-        console.error('CSRF token not found: https://laravel.com/docs/csrf#csrf-x-csrf-token');
-    }
-});
