+++ resources/js/elements/index.js
@@ -1,0 +1,2 @@
+import 'elements/turbo-echo-stream-tag';
+import 'elements/turbo-livewire-stream-tag';
