+++ resources/js/elements/turbo-echo-stream-tag.js
@@ -1,5 +1,4 @@
 import { connectStreamSource, disconnectStreamSource } from '@hotwired/turbo'
-import Echo from 'echo'
 
 const subscribeTo = (type, channel) => {
     if (type === "presence") {
