+++ resources/js/libs/index.js
@@ -1,0 +1,8 @@
+import 'libs/lodash';
+import 'libs/axios';
+import 'libs/stimulus';
+import 'libs/alpine';
+import 'libs/trix';
+import 'libs/turbo';
+import 'libs/bridge';
+import 'libs/echo';
