+++ resources/js/libs/lodash.js
@@ -1,0 +1,5 @@
+import _ from 'lodash';
+
+window._ = _;
+
+export default _;
